split SequentialModel -> Sequential, SequentialRNN
