+++ tutorials/learning_algorithms/zhang_debug.py
@@ -268,7 +268,7 @@ def zhang_train(data, model, **kwargs):
 
 if __name__ == '__main__':
 	curbd_data = np.load("data/ts/curbd_Adata.npy")
-	network = nt.SequentialModel(
+	network = nt.SequentialRNN(
 		layers=[nt.WilsonCowanLayer(
 			curbd_data.shape[0], curbd_data.shape[0],
 			activation="tanh",
